Banking Prototype v0.1.0 Account

Pure functional module that manages Account data structures

Operations are always registered on account's operations either if it is :denied or :ok

Types

t()

Specs

t() :: %Account{
  balances: map(),
  default_currency: atom(),
  limit: integer(),
  operations: map(),
  operations_auto_id: integer()
}

Basic struct that represents an Account

The Account.t() data structure is composed by 5 values:

balances: map containing %{currecy => balance} eg: %{BRL: 2500}

limit: The minimal balance required to make operations, this feature only works for account's default currency other currencies are always 0

operations: A map containing all the operations of a specific Account. eg %{1 => Account.Operation.t(), 2 => Account.Operation.t()}

default_currency: Account's default currency used on limit feature

operations_auto_id: Used internally to generate operations ids

Functions

balance(account, currency)

Specs

balance(t(), atom()) :: number()

Get the current balance of the account for the given currency

Examples

iex> init_account = Account.new(%{balances: %{BRL: 1000}})
iex> Account.balance(init_account, :BRL)
1000

balances(account)

Specs

balances(t()) :: map()

Get all the current balances of the account

Examples

iex> init_account = Account.new(%{balances: %{BRL: 1000, USD: 700, EUR: 300}})
iex> Account.balances(init_account)
%{BRL: 1000, USD: 700, EUR: 300}

card_transaction(account, data)

Specs

card_transaction(t(), %{
  amount: non_neg_integer(),
  currency: atom(),
  card_id: any()
}) ::
  {:ok, t(), Account.Operation.t()}
  | {:denied, String.t(), t(), Account.Operation.t()}

Register an event of card transaction and update the balance

Examples

iex> init_account = Account.new(%{balances: %{BRL: 1000}})
iex> {
...> :ok,
...> %Account{balances: %{BRL: 300}},
...> %Account.Operation{type: :card_transaction, data: %{amount: 700, currency: :BRL}}
...> } = Account.card_transaction(init_account, %{amount: 700, currency: :BRL, card_id: 1})

iex> init_account = Account.new(%{balances: %{BRL: -500}})
iex> {
...> :denied,
...> reason,
...> %Account{balances: %{BRL: -500}},
...> %Account.Operation{type: :card_transaction, data: %{amount: 700, currency: :BRL}}
...> } = Account.card_transaction(init_account, %{amount: 700, currency: :BRL, card_id: 1})

deposit(account, data)

Specs

deposit(t(), %{amount: pos_integer(), currency: atom()}) ::
  {:ok, t(), Account.Operation.t()}

Register an event of deposit and update the balance

Examples

iex> init_account = Account.new()
iex> {
...> :ok,
...> %Account{balances: %{BRL: 700}},
...> %Account.Operation{type: :deposit, data: %{amount: 700, currency: :BRL}}
...> } = Account.deposit(init_account, %{amount: 700, currency: :BRL})

exchange_balances(account, data)

Register an event of exchange, update the balances based on exchange rates

Examples

iex> init_account = Account.new(%{balances: %{USD: 1000}})
iex> {
...> :ok,
...> %Account{balances: %{BRL: 545, USD: 900}},
...> %Account.Operation{type: :exchange, status: :done}
...> } = Account.exchange_balances(init_account, %{current_amount: 100, current_currency: :USD, new_currency: :BRL})

new()

Specs

new() :: t()

Create a new Account data structure

Examples

iex> Account.new()
%Account{balances: %{BRL: 0}, limit: -500, operations: %{}, operations_auto_id: 1}

new(args)

Specs

new(map()) :: t()

Create a new Account data strucure with modified data

Examples

iex> entry_map = %{balances: %{BRL: 1000}, limit: -999}
iex> Account.new(entry_map)
%Account{balances: %{BRL: 1000}, limit: -999, operations: %{}, operations_auto_id: 1}

operation(account, operation_id)

Get the account's operation under the given id

Examples

iex> init_account = Account.new(%{balances: %{BRL: 1000}})
iex> {:ok, new_account, _} = Account.withdraw(init_account, %{amount: 700, currency: :BRL})
iex> %Account.Operation{} = Account.operation(new_account, 1)

operations(account, date)

Specs

operations(t(), Date.t()) :: [Account.Operation.t()]

Get a ordered list of all the operations that happen on a given date, ordered by occurence date time

Examples

iex> init_account = Account.new(%{balances: %{BRL: 1000}})
iex> {:ok, new_account, _} = Account.withdraw(init_account, %{amount: 700, currency: :BRL, date_time: ~U[2020-07-24 11:00:00Z]})
iex> {:denied, _, new_account, _} = Account.withdraw(new_account, %{amount: 1300, currency: :BRL, date_time: ~U[2020-07-24 12:00:00Z]})
iex> {:ok, new_account, _} = Account.deposit(new_account, %{amount: 700, currency: :BRL, date_time: ~U[2020-07-25 11:00:00Z]})
iex> [
...>  %Account.Operation{type: :withdraw, data: %{amount: 1300, currency: :BRL}, status: :denied},
...>  %Account.Operation{type: :withdraw, data: %{amount: 700, currency: :BRL}, status: :done}
...> ] = Account.operations(new_account, ~D[2020-07-24])

operations(account, ini_date, fin_date)

Specs

operations(t(), Date.t(), Date.t()) :: [Account.Operation.t()]

Get a ordered list of all the operations that happen between 2 dates, ordered by occurence date time

Examples

iex> init_account = Account.new(%{balances: %{BRL: 1000}})
iex> {:ok, new_account, _} = Account.withdraw(init_account, %{amount: 700, currency: :BRL, date_time: ~U[2020-07-24 11:00:00Z]})
iex> {:denied, _, new_account, _} = Account.withdraw(new_account, %{amount: 1300, currency: :BRL, date_time: ~U[2020-07-24 12:00:00Z]})
iex> {:ok, new_account, _} = Account.deposit(new_account, %{amount: 700, currency: :BRL, date_time: ~U[2020-07-25 11:00:00Z]})
iex> {:ok, new_account, _} = Account.deposit(new_account, %{amount: 1800, currency: :BRL, date_time: ~U[2020-07-26 11:00:00Z]})
iex> [
...>  %Account.Operation{type: :deposit, data: %{amount: 700, currency: :BRL}, status: :done},
...>  %Account.Operation{type: :withdraw, data: %{amount: 1300, currency: :BRL}, status: :denied},
...>  %Account.Operation{type: :withdraw, data: %{amount: 700, currency: :BRL}, status: :done}
...> ] = Account.operations(new_account, ~D[2020-07-24], ~D[2020-07-25])

refund(account, data)

Specs

refund(t(), %{operation_to_refund_id: any()}) ::
  {:ok, t(), Account.Operation.t()} | {:error, String.t(), t()}

Register an event of refund, update de balance and update the refunded operation status

Only card operations are refundable, all other operations will return {:error, reason, account_data}

Examples

iex> init_account = Account.new(%{balances: %{BRL: 1000}})
iex> {:ok, init_account, %{id: op_id}} = Account.card_transaction(init_account, %{amount: 700,currency: :BRL, card_id: 1})
iex> {
...> :ok,
...> %Account{balances: %{BRL: 1000}},
...> %Account.Operation{type: :refund, data: %{operation_to_refund_id: ^op_id}}
...> } = Account.refund(init_account, %{operation_to_refund_id: op_id})

transfer_in(account, data)

Specs

transfer_in(t(), %{
  amount: pos_integer(),
  currency: atom(),
  sender_account_id: any()
}) :: {:ok, t(), Account.Operation.t()}

Register an event of transfer in and update the balance

Examples

iex> init_account = Account.new(%{balances: %{BRL: 300}})
iex> {
...> :ok,
...> %Account{balances: %{BRL: 1000}},
...> %Account.Operation{type: :transfer_in, data: %{amount: 700, currency: :BRL, sender_account_id: 1}}
...> } = Account.transfer_in(init_account, %{amount: 700, currency: :BRL, sender_account_id: 1})

transfer_out(account, data)

Specs

transfer_out(t(), %{
  amount: non_neg_integer(),
  currency: atom(),
  recipient_account_id: number()
}) ::
  {:ok, t(), Account.Operation.t()}
  | {:denied, String.t(), t(), Account.Operation.t()}

transfer_out(t(), %{amount: non_neg_integer(), recipients_data: list()}) ::
  {:ok, t(), [Account.Operation.t()]}
  | {:denied, String.t(), t(), Account.Operation.t()}

Register an event of transfer_out for each data received on the list and update the balance

If suceeds, the split operation generates N :transfer_out operations on the account operation list

If it is denied, only one operation will be created on the operations lists

All the aditional data passed to data paramenter will be copied to each generated operation

Examples

iex> init_account = Account.new(%{balances: %{BRL: 1000}})
iex> {
...> :ok,
...> %Account{balances: %{BRL: 300}},
...> %Account.Operation{type: :transfer_out, data: %{amount: 700, currency: :BRL}}
...> } = Account.transfer_out(init_account, %{amount: 700, recipient_account_id: 1, currency: :BRL})

iex> init_account = Account.new(%{balances: %{BRL: -500}})
iex> {
...> :denied,
...> reason,
...> %Account{balances: %{BRL: -500}},
...> %Account.Operation{type: :transfer_out, data: %{amount: 700, currency: :BRL}}
...> } = Account.transfer_out(init_account, %{amount: 700, currency: :BRL, recipient_account_id: 1})

iex> init_account = Account.new(%{balances: %{BRL: 3000}})
iex> data = %{
...>  currency: :BRL,
...>  amount: 1000,
...>  recipients_data: [
...>    %{percentage: 0.7, recipient_account_id: 2},
...>    %{percentage: 0.2, recipient_account_id: 3},
...>    %{percentage: 0.1, recipient_account_id: 4}
...>  ]}
iex>  {
...>  :ok,
...>  %Account{balances: %{BRL: 2000}},
...>  [
...>    %Account.Operation{data: %{amount: 100, currency: :BRL, recipient_account_id: 4}},
...>    %Account.Operation{data: %{amount: 200, currency: :BRL, recipient_account_id: 3}},
...>    %Account.Operation{data: %{amount: 700, currency: :BRL, recipient_account_id: 2}}
...> ]} = Account.transfer_out(init_account, data)

withdraw(account, data)

Specs

withdraw(t(), %{amount: non_neg_integer(), currency: atom()}) ::
  {:ok, t(), Account.Operation.t()}
  | {:denied, String.t(), t(), Account.Operation.t()}

Register an event of withdraw and update account's balance

Examples

iex> init_account = Account.new(%{balances: %{BRL: 1000}})
iex> {
...> :ok,
...> %Account{balances: %{BRL: 300}},
...> %Account.Operation{type: :withdraw, data: %{amount: 700, currency: :BRL}}
...> } = Account.withdraw(init_account, %{amount: 700, currency: :BRL})

iex> init_account = Account.new(%{balances: %{BRL: -500}})
iex> {
...> :denied,
...> reason,
...> %Account{balances: %{BRL: -500}},
...> %Account.Operation{type: :withdraw, data: %{amount: 50, currency: :BRL}}
...> } = Account.withdraw(init_account, %{amount: 50, currency: :BRL})