Banking Prototype v0.1.0 Account.Cache

DynamicSupervisor that manages all Account.Server processes currently running on the system

This module is used as an Account.Server dicover, always that some part of the system needs to issue requests to an Account.Server, it first asks if some process containing the data of a specific account is already running

Functions

run_server_process(account_id, args \\ %{})

Get the pid of a Account.Server process that is running with the account's data that has the given id.

If the server isn't running it is initialized with the data persisted on database. If no data is already persisted, the server is initialized with the given args

Examples

iex> bob_account_pid = Account.Cache.server_process(1)
iex> is_pid(bob_account_pid)
true

server_process(account_id, args \\ %{})