Banking Prototype v0.1.0 Account.Operation

Pure functional module, used to define new operations that happen over an Account data structure

By default, any operation is initialized with status :done and with the current date time

If you want to change any key value of the operation struct you should pass it new value inside the data parameter eg: Account.Operation.new(:type, %{status: :new_status} )

Types

t()

Specs

t() :: %Account.Operation{
  data: map(),
  date_time: Date.t(),
  id: term(),
  status: term(),
  type: atom()
}

Basic struct to define an Account.Operation

The Account.Operation.t() data structure is composed by 5 values:

date_time: The date and time of the operation occurence

type: Atom that identifies the type of the operation

data: Customized data about the operation based on the operation type, can be used to inject metadata about the operation

status: Atom that indicates if the operation suceeded or not. [:done, :denied, :refunded]

id: Sequential identifier, that identifies the operation only inside an Account context

Functions

new(atom, entry_data)

Specs

new(:withdraw, %{amount: number()} | map()) :: t()

new(:deposit, %{amount: number()} | map()) :: t()

new(:transfer_out, %{amount: number(), recipient_account_id: number()} | map()) ::
  t()

new(:transfer_in, %{amount: number(), sender_account_id: number()} | map()) ::
  t()

new(:card_transaction, %{amount: number(), card_id: number()} | map()) :: t()

new(:refund, %{operation_to_refund_id: number()} | map()) :: t()

new(
  :exchange,
  %{
    current_amount: number(),
    current_currency: atom(),
    new_amount: number(),
    new_currency: atom(),
    exchange_rate: number()
  }
  | map()
) :: t()

new(
  :exchange,
  %{
    current_amount: number(),
    current_currency: atom(),
    new_currency: atom(),
    status: :denied
  }
  | map()
) :: t()

Create a new operation based on the parameters

Any extra data passed on the data parameter will be part of the final term, inside data key

You can pass key arguments on data to change operation default values such as :date_time and :status

Examples

iex> oop = Account.Operation.new(:withdraw, %{amount: 1000})
iex> match?(%{date_time: _, type: :withdraw, data: %{amount: 1000}, status: :done}, oop)
true

#Any extra data passed on the data parameter will be part of the final term
iex> oop = Account.Operation.new(:deposit, %{amount: 1000, meta_data: "some data"})
iex> match?(%{data: %{amount: 1000, meta_data: "some data"}}, oop)
true

# Any key argument on data will change the operation default value for the provided key
iex> oop = Account.Operation.new(:deposit, %{amount: 1000, status: "My custom status", date_time: ~U[2020-07-24 10:00:00Z]})
iex> match?(%{status: "My custom status", date_time: ~U[2020-07-24 10:00:00Z]}, oop)
true