Banking Prototype v0.1.0 Account.Server

GenServer that keeps track of an Account data structure as it's state

Functions

account_id(account_server)

Specs

account_id(pid()) :: number()

Get the account id of a specific account

Examples

iex> server_pid = Account.Cache.server_process(123)
iex> Account.Server.account_id(server_pid)
123

balance(account_server, currency)

Specs

balance(pid(), atom()) :: number()

Get the current balance of the Account

Examples

iex> server_pid = Account.Cache.server_process(1, %{balances: %{BRL: 3000}})
iex> Account.Server.balance(server_pid, :BRL)
3000

balances(account_server)

card_transaction(account_server, data)

Specs

card_transaction(pid(), %{amount: number(), currency: atom(), card_id: number()}) ::
  {:ok, number(), Account.Operation.t()}
  | {:denied, String.t(), number(), Account.Operation.t()}

Debit card operation that uses resources from the account balances

Examples

iex> server_pid = Account.Cache.server_process(1, %{balances: %{BRL: 3000}})
iex> {
...>  :ok,
...>  1000,
...>  %Account.Operation{type: :card_transaction, status: :done, data: %{amount: 2000, currency: :BRL, card_id: 1}},
...> } = Account.Server.card_transaction(server_pid, %{amount: 2000, currency: :BRL, card_id: 1})

iex> server_pid = Account.Cache.server_process(1, %{balances: %{BRL: 500}})
iex> {
...>  :denied,
...>  reason,
...>  500,
...>  %Account.Operation{type: :card_transaction, status: :denied, data: %{amount: 2000, currency: :BRL}}
...> } = Account.Server.card_transaction(server_pid, %{amount: 2000, currency: :BRL, card_id: 1})

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

deposit(account_server, data)

Specs

deposit(pid(), %{amount: non_neg_integer(), currency: atom()}) ::
  {:ok, number(), Account.Operation.t()}

Deposit into the account balance

Examples

iex> server_pid = Account.Cache.server_process(1, %{balances: %{BRL: 3000}})
iex> {
...>  :ok,
...>  5000,
...>  %Account.Operation{type: :deposit, status: :done, data: %{amount: 2000, currency: :BRL}}
...> } = Account.Server.deposit(server_pid, %{amount: 2000, currency: :BRL})

exchange_balances(account_server, data)

Register an exchange operation and update the balances using the current exchange rate

Examples

iex> server_pid = Account.Cache.server_process(1, %{balances: %{BRL: 3000}})
iex> {
...>  :ok,
...>  %{BRL: 2000, USD: 183},
...>  %Account.Operation{type: :exchange, status: :done},
...> } = Account.Server.exchange_balances(server_pid, %{current_amount: 1000, current_currency: :BRL, new_currency: :USD})

operation(account_server, operation_id)

Specs

operation(pid(), number()) :: Account.Operation.t() | nil

Get operation data under the given id

Examples

iex> server_pid = Account.Cache.server_process(123)
iex> Account.Server.deposit(server_pid, %{amount: 1000, currency: :BRL})
iex> %Account.Operation{
...>   type: :deposit,
...>   data: %{amount: 1000, currency: :BRL},
...>   status: :done
...> } = Account.Server.operation(server_pid, 1)

operations(account_server, date)

Specs

operations(pid(), Date.t()) :: [Account.Operation.t()]

Get the list of operations that occurred in a specific date

Examples

iex> server_pid = Account.Cache.server_process(1, %{balances: %{BRL: 2000}})
iex> Account.Server.withdraw(server_pid, %{amount: 300, currency: :BRL, date_time: ~U[2020-07-23 10:00:00Z]})
iex> Account.Server.withdraw(server_pid, %{amount: 700, currency: :BRL, date_time: ~U[2020-07-24 11:00:00Z]})
iex> Account.Server.withdraw(server_pid, %{amount: 1900, currency: :BRL, date_time: ~U[2020-07-24 12:00:00Z]})
iex> [
...>  %Account.Operation{type: :withdraw, data: %{amount: 1900, currency: :BRL}, status: :denied},
...>  %Account.Operation{type: :withdraw, data: %{amount: 700, currency: :BRL}, status: :done}
...> ] = Account.Server.operations(server_pid, ~D[2020-07-24])

operations(account_server, date_ini, date_fin)

Specs

operations(pid(), Date.t(), Date.t()) :: [Account.Operation.t()]

Get the list of operations that occur betweem 2 dates, ordered by occurence date time

Examples

iex> server_pid = Account.Cache.server_process(1, %{balances: %{BRL: 2000}})
iex> Account.Server.withdraw(server_pid, %{amount: 300, currency: :BRL, date_time: ~U[2020-07-23 10:00:00Z]})
iex> Account.Server.withdraw(server_pid, %{amount: 700, currency: :BRL, date_time: ~U[2020-07-24 11:00:00Z]})
iex> Account.Server.withdraw(server_pid, %{amount: 1900, currency: :BRL, date_time: ~U[2020-07-25 12:00:00Z]})
iex> [
...>  %Account.Operation{type: :withdraw, data: %{amount: 700, currency: :BRL}, status: :done},
...>  %Account.Operation{type: :withdraw, data: %{amount: 300, currency: :BRL}, status: :done}
...> ] = Account.Server.operations(server_pid, ~D[2020-07-23], ~D[2020-07-24])

refund(account_server, data)

Specs

refund(pid(), %{operation_to_refund_id: any()}) ::
  {:ok, map(), Account.Operation.t()}
  | {:denied, String.t(), number(), Account.Operation.t()}
  | {:error, String.t(), Account.Operation.t()}

Register a refund operation, restore the amount to balance and update operation status to :refunded

Examples

iex> server_pid = Account.Cache.server_process(1, %{balances: %{BRL: 3000}})
iex> Account.Server.card_transaction(server_pid, %{amount: 2000, currency: :BRL, card_id: 1})
iex> {
...>  :ok,
...>  %{BRL: 3000},
...>  %Account.Operation{type: :refund, status: :done, data: %{operation_to_refund_id: 1}},
...> } = Account.Server.refund(server_pid, %{operation_to_refund_id: 1})

start_link(args)

Specs

start_link(%{id: number()}) :: :ignore | {:error, any()} | {:ok, pid()}

Start the Account.Server server

Should not be called directly, instead use Account.Cache.server_process/2

The initial state of non persisted non args Account.Server processes is the result of Account.new/0

transfer_in(account_server, data)

Specs

transfer_in(pid(), %{
  amount: non_neg_integer(),
  currency: atom(),
  sender_account_id: number()
}) :: {:ok, number(), Account.Operation.t()}

Deposit into the account balance

Examples

iex> server_pid = Account.Cache.server_process(1, %{balances: %{BRL: 3000}})
iex> {
...>  :ok,
...>  5000,
...>  %Account.Operation{type: :transfer_in, status: :done, data: %{amount: 2000, currency: :BRL}}
...> } = Account.Server.transfer_in(server_pid, %{amount: 2000, currency: :BRL, sender_account_id: 2})

transfer_out(account_server, data)

Specs

transfer_out(pid(), %{
  amount: non_neg_integer(),
  currency: atom(),
  recipient_account_id: number()
}) ::
  {:ok, number(), number(), Account.Operation.t(), Account.Operation.t()}
  | {:denied, String.t(), number(), Account.Operation.t()}

transfer_out(
  pid(),
  %{amount: number(), currency: atom(), recipients_data: list()} | map()
) ::
  {:ok, number(), number(), [Account.Operation.t()], [Account.Operation.t()]}
  | {:denied, String.t(), number(), Account.Operation.t()}

Transfer resources to another accounts

The version of this functions that receives recipients_data make several operations according to the given list length

Although the list version relies heavily on the simple single recipient implementation, the list call is just syntatic sugar to make multiple single transfers

In case of success there is no differece between multiple single transfers and a single multi tranfer, it is impossible to know after the fact

The only difference is in case of denial, because the multi transfer will generate a single operation with all the data, instead of multiple transfers

Examples

iex> server_pid = Account.Cache.server_process(1, %{balances: %{BRL: 3000}})
iex> {
...>  :ok,
...>  1000,
...>  %Account.Operation{type: :transfer_out, status: :done, data: %{amount: 2000,currency: :BRL, recipient_account_id: 2}},
...>  %Account.Operation{type: :transfer_in, status: :done, data: %{amount: 2000,currency: :BRL, sender_account_id: 1}}
...> } = Account.Server.transfer_out(server_pid, %{amount: 2000, currency: :BRL, recipient_account_id: 2})

iex> server_pid = Account.Cache.server_process(1, %{balances: %{BRL: 500}})
iex> {
...>  :denied,
...>  reason,
...>  500,
...>  %Account.Operation{type: :transfer_out, status: :denied, data: %{amount: 2000, currency: :BRL}}
...> } = Account.Server.transfer_out(server_pid, %{amount: 2000, currency: :BRL, recipient_account_id: 2})

iex> server_pid = Account.Cache.server_process(1, %{balances: %{BRL: 500}})
iex> recipients_data = [
...>   %{percentage: 0.7, recipient_account_id: 2, other_data: "another extra data"},
...>   %{percentage: 0.2, recipient_account_id: 3, meta_data: "specific meta_data"},
...>   %{percentage: 0.1, recipient_account_id: 4}
...> ]
iex> {
...>  :denied,
...>  reason,
...>  500,
...>  %Account.Operation{type: :transfer_out, status: :denied, data: %{amount: 2000, currency: :BRL}}
...> } = Account.Server.transfer_out(server_pid, %{amount: 2000, currency: :BRL, recipients_data: recipients_data})

withdraw(account_server, data)

Specs

withdraw(pid(), %{amount: non_neg_integer(), currency: atom()}) ::
  {:ok, number(), Account.Operation.t()}
  | {:denied, String.t(), number(), Account.Operation.t()}

Withdraw from the account balance

Examples

iex> server_pid = Account.Cache.server_process(1, %{balances: %{BRL: 3000}})
iex> {
...>  :ok,
...>  1000,
...>  %Account.Operation{type: :withdraw, status: :done, data: %{amount: 2000, currency: :BRL}}
...> } = Account.Server.withdraw(server_pid, %{amount: 2000, currency: :BRL})

iex> server_pid = Account.Cache.server_process(1, %{balances: %{BRL: 500}})
iex> {
...>  :denied,
...>  reason,
...>  500,
...>  %Account.Operation{type: :withdraw, status: :denied, data: %{amount: 2000, currency: :BRL}}
...> } = Account.Server.withdraw(server_pid, %{amount: 2000, currency: :BRL})