Banking Prototype v0.1.0 Database

Supervisor that manages Database.Worker processes, responsible for data persistence functionality and workers pooling

Database.worker are used to perform kind of connection pools, spliting the real computational work between all processes

This module works as a connection pool manager, just forwarding the request for a selected worker

This module guarantee that requests that manipulate the same key will always be handled by the same Database.worker process, resulting in no race conditions, because worker processes receives messages sequentially

Internally, all the real computation is on Database.Worker processes

Functions

child_spec(_)

get(key, folder)

Specs

get(any(), String.t()) :: any() | nil

Get the persisted data registered under the given folder with the given key

start_link()

Specs

start_link() :: :ignore | {:error, any()} | {:ok, pid()}

Start the server

store_async(key, value, folder)

store_local_async(key, value, folder)

Specs

store_local_async(any(), any(), String.t()) :: :ok

Persist data asynchronously under the given folder with the given key

store_local_sync(key, value, folder)

Specs

store_local_sync(any(), any(), String.t()) :: :ok

Persist data synchronously under the given folder with the given key

store_sync(key, value, folder)