Banking Prototype v0.1.0 Database.Worker

GenServer that handle data persistance requests

This module is the bottle neck of the system, since the implementation is really naive, and relies just on binary data and the file system, the throughput is not optimal

This module is responsible for 60% of the response time

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

get(worker_id, key, folder)

Specs

get(number(), any(), String.t()) :: any()

Get the persisted data registered under the given folder with the given key

init(_)

Callback implementation for GenServer.init/1.

start_link(worker_id)

Specs

start_link({number(), String.t()}) :: :ignore | {:error, any()} | {:ok, pid()}

store_async(worker_id, key, data, folder)

Specs

store_async(number(), any(), any(), String.t()) :: :ok

Persist data asynchronously under the given folder with the given key

store_sync(worker_id, key, data, folder)

Specs

store_sync(number(), any(), any(), String.t()) :: any()

Persist data synchronously under the given folder with the given key